Save refactor_push state
