+++ package.json
@@ -49,8 +49,8 @@
     "fs": "^0.0.1-security",
     "fuzzy": "^0.1.3",
     "globby": "^9.0.0",
-    "google-auth-library": "^2.0.0",
-    "googleapis": "^36.0.0",
+    "google-auth-library": "^3.0.1",
+    "googleapis": "^37.1.0",
     "inquirer": "^5.2.0",
     "inquirer-autocomplete-prompt": "1.0.1",
     "is-online": "^7.0.0",
