+++ src/commands/help.ts
@@ -1,0 +1,9 @@
+import * as commander from 'commander';
+
+/**
+ * Outputs the help command.
+ */
+export default async () => {
+  commander.outputHelp();
+  process.exit(0);
+};
