+++ src/fileutils.ts
@@ -3,8 +3,8 @@ import * as path from 'path';
 import * as recursive from 'recursive-readdir';
 import * as ts from 'typescript';
 import {DOT} from './dotfile';
-import { ProjectFile } from './fileutils';
-import { ERROR, getAPIFileType, logError, spinner } from './utils';
+import { getFileType } from './files';
+import { ERROR, getProjectSettings, logError, spinner } from './utils';
 const readMultipleFiles = require('read-multiple-files');
 const ts2gas = require('ts2gas');
 const findParentDir = require('find-parent-dir');
@@ -16,17 +16,17 @@ export interface AppsScriptAPIFile {
   source: string;
 }
 
-// A file that may be ignored by the project.
-export interface ProjectFile {
-  fileContents: AppsScriptAPIFile;
-  ignored: boolean;
+// A generic file interface.
+export interface File {
+  name: string;
+  content: string;
 }
 
 /**
  * Gets a list of all files that are within the specified directory's folder.
  * @param {string} rootDir The base directory to get the file paths at.
  */
-export async function getFilePaths(rootDir: string): Promise<string[]> {
+export async function getAllFilePaths(rootDir: string): Promise<string[]> {
   return new Promise<string[]>((res, rej) => {
     recursive(rootDir, async (err, filePaths) => {
       if (err) return rej(err);
@@ -36,16 +36,49 @@ export async function getFilePaths(rootDir: string): Promise<string[]> {
 }
 
 /**
+ * Filters out bad file paths.
+ * Valid paths includes .ts, .gs, .html files.
+ * Valid paths includes appsscript.json file.
+ */
+export async function getValidFilePaths(filePaths: string[]): Promise<string[]> {
+  return filePaths.filter(filePath => {
+    const ext = path.extname(filePath).toLowerCase();
+    const isJSFile = ext === '.js' || ext === '.ts' || ext === '.gs';
+    const isManifest = filePath === 'appsscript.json';
+    // console.log(isJSFile, isManifest, filePath);
+    return isJSFile || isManifest; // don't filter these files.
+  });
+}
+
+/**
  * Gets the contents of a list of files.
  * @param {string[]} filePaths A list of file paths.
  */
-export async function getFileContents(filePaths: string[]): Promise<string[]> {
-  return new Promise<string[]>((res, rej) => {
+export async function getFileContents(filePaths: string[]): Promise<File[]> {
+  return new Promise<File[]>((res, rej) => {
     readMultipleFiles(filePaths, 'utf8', (err: string, contents: string[]) => {
       if (err) return rej(err);
-      res(contents);
+      const o = contents.map((content: string, i: number) => {
+        return {
+          name: filePaths[i],
+          content,
+        };
+      });
+      res(o);
     });
   });
+}
+
+/**
+ * Fixes OS specific path separator.
+ * @param {Files[]} files The files to fix the path.
+ */
+export async function fixFilePaths(files: File[]): Promise<File[]> {
+  files.map(f => {
+    // Replace OS specific path separator to common '/' char for console output
+    f.name = f.name.replace(/\\/g, '/');
+  });
+  return files;
 }
 
 /**
@@ -77,81 +110,12 @@ export async function checkConflictingFilePaths(filePaths: string[]): Promise<bo
 }
 
 /**
- * Gets all files:
- * - Non-ignored files: Get the name and contents
- * - Ignored files: Get the name
- */
-export async function getAllFiles({
-  filePaths,
-  fileContents,
-  ignoredFilePaths,
-  nonIgnoredFilePaths,
-  ignoreMatches,
-  rootDir,
-}: {
-  filePaths: string[],
-  fileContents: string[],
-  ignoredFilePaths: string[],
-  ignoreMatches: string[],
-  nonIgnoredFilePaths: string[],
-  rootDir: string,
-}): Promise<ProjectFile[]> {
-  const files: Array<AppsScriptAPIFile | undefined> = filePaths.map((name, i) => {
-    const normalizedName = path.normalize(name);
-    const type = getAPIFileType(name);
-
-    // Formats rootDir/appsscript.json to appsscript.json.
-    // Preserves subdirectory names in rootDir
-    // (rootDir/foo/Code.js becomes foo/Code.js)
-    const formattedName = getAppsScriptFileName(rootDir, name);
-
-    // If the file is valid, return the file in a format suited for the Apps Script API.
-    if (isValidFileName({
-      name,
-      type,
-      rootDir,
-      normalizedName,
-      ignoreMatches,
-    })) {
-      nonIgnoredFilePaths.push(name);
-      const file: AppsScriptAPIFile = {
-        name: formattedName, // the file base name
-        type, // the file extension
-        source: fileContents[i], //the file contents
-      };
-      return file;
-    } else {
-      ignoredFilePaths.push(name);
-      return undefined; // Skip ignored files
-    }
-  });
-  // Fixes TypeScript errors.
-  function notEmpty<T>(value: T | undefined): value is T {
-    return !!value;
-  }
-  const nonNullFiles: AppsScriptAPIFile[] = files.filter(notEmpty);
-  const projectFilesList: ProjectFile[] = nonNullFiles.map((f: AppsScriptAPIFile) => {
-    const projectFile: ProjectFile = {
-      fileContents: {
-        name: '',
-        type: '',
-        source: '',
-      },
-      ignored: false,
-    };
-    return projectFile;
-  });
-  return projectFilesList;
-}
-
-/**
  * Returns in tsconfig.json.
  * @returns {ts.TranspileOptions} if tsconfig.json not exists, return undefined.
  */
 export function getTranspileOptions(): ts.TranspileOptions {
   const projectDirectory: string = findParentDir.sync(process.cwd(), DOT.PROJECT.PATH) || DOT.PROJECT.DIR;
   const tsconfigPath = path.join(projectDirectory, 'tsconfig.json');
-  const userConf: ts.TranspileOptions = {};
   if (fs.existsSync(tsconfigPath)) {
     const tsconfigContent = fs.readFileSync(tsconfigPath, 'utf8');
     const parsedConfigResult = ts.parseConfigFileTextToJson(tsconfigPath, tsconfigContent);
@@ -180,19 +144,34 @@ export function getAppsScriptFileName(rootDir: string, filePath: string): string
 
 /**
  * Transpiles TS files to GS
- * @param {ProjectFile[]} projectFiles A list of project files.
+ * @param {ProjectFile[]} files A list of project files.
  * @see http://github.com/grant/ts2gas
  */
-export function transpileTsFiles(projectFiles: ProjectFile[]): ProjectFile[] {
+export async function transpileTsFiles(files: File[]): Promise<File[]> {
   // Load tsconfig
   const userTranspileOptions: ts.TranspileOptions = getTranspileOptions();
-  return projectFiles.map((projectFile: ProjectFile) => {
+  return files.map((file: File) => {
     // File source
-    const source = projectFile.fileContents.source;
-    if (projectFile.fileContents.type === 'TS') {
-      projectFile.fileContents.source = ts2gas(source, userTranspileOptions);
+    const source = file.content;
+    if (path.extname(file.name).toUpperCase() === '.TS') {
+      file.content = ts2gas(source, userTranspileOptions);
     }
-    return projectFile;
+    return file;
+  });
+}
+
+/**
+ * Removes the extension from the name.
+ */
+export async function removeExtensionFromName(files: File[]): Promise<File[]> {
+  return files.map((file: File) => {
+    const ext = path.extname(file.name);
+    const fileNameWithoutExt = file.name.slice(0, -ext.length);
+    return {
+      name: fileNameWithoutExt,
+      content: file.content,
+      type: getFileType(ext),
+    };
   });
 }
 
@@ -200,62 +179,26 @@ export function transpileTsFiles(projectFiles: ProjectFile[]): ProjectFile[] {
  * Sorts files by push order.
  * @param {ProjectFile[]} ProjectFile A list of project files.
  */
-export const sortFilesByPushOrder = (files: ProjectFile[], filePushOrder: string[]) => {
+export async function sortFilesByPushOrder(files: File[]): Promise<File[]> {
   // This statement customizes the order in which the files are pushed.
   // It puts the files in the setting's filePushOrder first.
   // This is needed because Apps Script blindly executes files in order of creation time.
   // The Apps Script API updates the creation time of files.
-  spinner.stop(true);
+  const { filePushOrder } = await getProjectSettings();
+  if (!filePushOrder) return files;
   console.log('Detected filePushOrder setting. Pushing these files first:');
   filePushOrder.map(file => {
     console.log(`└─ ${file}`);
   });
   console.log('');
-  return files = files.sort((file1: ProjectFile, file2: ProjectFile) => {
+  return files = files.sort((file1: File, file2: File) => {
+    if (!filePushOrder) return 0; // TS complains
     // Get the file order index
-    let path1Index = filePushOrder.indexOf(file1.fileContents.name);
-    let path2Index = filePushOrder.indexOf(file2.fileContents.name);
+    let path1Index = filePushOrder.indexOf(file1.name);
+    let path2Index = filePushOrder.indexOf(file2.name);
     // If a file path isn't in the filePushOrder array, set the order to -∞.
     path1Index = path1Index === -1 ? Number.NEGATIVE_INFINITY : path1Index;
     path2Index = path2Index === -1 ? Number.NEGATIVE_INFINITY : path2Index;
     return path2Index - path1Index;
   });
-};
-
-/**
- * If the file is valid, add it to our file list.
- * We generally want to allow for all file types, including files in node_modules/.
- * However, node_modules/@types/ files should be ignored.
- */
-const isValidFileName = ({
-  name,
-  type,
-  rootDir,
-  normalizedName,
-  ignoreMatches,
-}: {
-  name: string,
-  type: string,
-  rootDir: string,
-  normalizedName: string,
-  ignoreMatches: string[],
-}) => {
-  let valid = true; // Valid by default, until proven otherwise.
-  // Has a type or is appsscript.json
-  let isValidJSONIfJSON = true;
-  if (type === 'JSON') {
-    if (rootDir) {
-      isValidJSONIfJSON = normalizedName === path.join(rootDir, 'appsscript.json');
-    } else {
-      isValidJSONIfJSON = name === 'appsscript.json';
-    }
-  } else {
-    // Must be SERVER_JS or HTML.
-    // https://developers.google.com/apps-script/api/reference/rest/v1/File
-    valid = type === 'SERVER_JS' || type === 'HTML';
-  }
-  const validType = type && isValidJSONIfJSON;
-  const notIgnored = !ignoreMatches.includes(name);
-  valid = !!(valid && validType && notIgnored);
-  return valid;
-};
+}
