+++ src/index.ts
@@ -23,28 +23,26 @@
 import 'connect';
 
 import * as commander from 'commander';
-import {
-  apis,
-  defaultCmd,
-  deploy,
-  help,
-  list,
-  login,
-  logout,
-  logs,
-  push,
-  run,
-  status,
-  undeploy,
-  versions,
-} from './commands';
+import apis from './commands/apis';
 import clone from './commands/clone';
 import create from './commands/create';
+import defaultCmd from './commands/defaultCmd';
+import deploy from './commands/deploy';
 import deployments from './commands/deployments';
+import help from './commands/help';
+import list from './commands/list';
+import login from './commands/login';
+import logout from './commands/logout';
+import logs from './commands/logs';
 import openCmd from './commands/open';
 import pull from './commands/pull';
+import push from './commands/push';
+import run from './commands/run';
 import setting from './commands/settings';
+import status from './commands/status';
+import undeploy from './commands/undeploy';
 import version from './commands/version';
+import versions from './commands/versions';
 import { PROJECT_NAME, handleError } from './utils';
 
 // const commander = require('commander');
